# specs/001-starter-app-structure/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Starter Web Application API
  description: |
    FastAPI backend for starter web application demonstrating authentication,
    user management, and best practices for domain-driven design.
    
    ## Authentication
    
    This API uses JWT bearer token authentication:
    1. Register or login to receive access and refresh tokens
    2. Include access token in Authorization header: `Bearer <token>`
    3. Access tokens expire after 15 minutes
    4. Use refresh token endpoint to get new access tokens
    5. Refresh tokens expire after 7 days
    
    ## Response Codes
    
    - 200: Success
    - 201: Created
    - 400: Bad Request (validation error)
    - 401: Unauthorized (invalid or missing token)
    - 403: Forbidden (insufficient permissions)
    - 404: Not Found
    - 409: Conflict (duplicate resource)
    - 422: Unprocessable Entity (validation error)
    - 500: Internal Server Error
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: auth
    description: Authentication and authorization operations
  - name: users
    description: User management operations
  - name: health
    description: Health check and system status

paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      description: Check if the API is running and responsive
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: "healthy"
                timestamp: "2025-10-24T12:00:00Z"
                version: "1.0.0"

  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register new user
      description: Create a new user account with email and password
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
            example:
              email: "user@example.com"
              password: "SecurePassword123!"
              full_name: "John Doe"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "CONFLICT"
                message: "Email already registered"

  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: User login
      description: Authenticate user and receive JWT tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
            example:
              email: "user@example.com"
              password: "SecurePassword123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              example:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                token_type: "bearer"
                expires_in: 900
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "UNAUTHORIZED"
                message: "Invalid email or password"

  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      description: Get a new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
            example:
              refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: User logout
      description: Revoke refresh token and logout user
      operationId: logoutUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully logged out"
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/users/me:
    get:
      tags: [users]
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [users]
      summary: Update current user profile
      description: Update the authenticated user's profile information
      operationId: updateCurrentUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
            example:
              full_name: "Jane Doe"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/users/me/password:
    put:
      tags: [users]
      summary: Change password
      description: Change the authenticated user's password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
            example:
              current_password: "OldPassword123!"
              new_password: "NewSecurePassword456!"
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password changed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or refresh endpoints

  schemas:
    # Auth Schemas
    UserRegistration:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address (used for login)
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          description: User's password (min 8 characters)
          example: "SecurePassword123!"
        full_name:
          type: string
          minLength: 2
          maxLength: 100
          description: User's full name
          example: "John Doe"

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          format: password
          description: User's password
          example: "SecurePassword123!"

    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token (expires in 15 minutes)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          description: JWT refresh token (expires in 7 days)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          enum: [bearer]
          description: Token type (always "bearer")
          example: "bearer"
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 900

    TokenRefresh:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Valid refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    # User Schemas
    UserResponse:
      type: object
      required:
        - id
        - email
        - full_name
        - is_active
        - is_verified
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique user identifier
          example: 1
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        full_name:
          type: string
          description: User's full name
          example: "John Doe"
        is_active:
          type: boolean
          description: Account active status
          example: true
        is_verified:
          type: boolean
          description: Email verification status
          example: true
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2025-10-24T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last profile update timestamp
          example: "2025-10-24T13:30:00Z"
        last_login_at:
          type: string
          format: date-time
          nullable: true
          description: Most recent login timestamp
          example: "2025-10-24T14:00:00Z"

    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 100
          description: Updated full name
          example: "Jane Doe"

    PasswordChange:
      type: object
      required:
        - current_password
        - new_password
      properties:
        current_password:
          type: string
          format: password
          description: Current password for verification
          example: "OldPassword123!"
        new_password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          description: New password (min 8 characters)
          example: "NewSecurePassword456!"

    # Common Schemas
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - version
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: API health status
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          description: Current server time
          example: "2025-10-24T12:00:00Z"
        version:
          type: string
          description: API version
          example: "1.0.0"

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Email is required"
        details:
          type: object
          description: Additional error details (validation errors, etc.)
          additionalProperties: true
          example:
            field: "email"
            constraint: "required"

    ValidationError:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            type: string
          description: Location of the error (path to field)
          example: ["body", "email"]
        msg:
          type: string
          description: Error message
          example: "field required"
        type:
          type: string
          description: Error type
          example: "value_error.missing"

  responses:
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  $ref: '#/components/schemas/ValidationError'
          example:
            detail:
              - loc: ["body", "email"]
                msg: "field required"
                type: "value_error.missing"

    Unauthorized:
      description: Unauthorized (missing or invalid token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Invalid or expired token"

    Forbidden:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "FORBIDDEN"
            message: "Insufficient permissions"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "NOT_FOUND"
            message: "Resource not found"

security:
  - BearerAuth: []